fix: extract getInitialThemeMode to testable module, fix ESM/TS issues
- Moves getInitialThemeMode to getInitialThemeMode.ts (no heavy deps)
- Removes inline function and ThemeMode import from EmbeddedContextProviders.tsx
- Replaces EmbeddedContextProviders.test.tsx with getInitialThemeMode.test.ts
  testing the pure function directly (avoids ESM/Jest issues from query-string)
- Uses getCurrentMode() instead of non-existent .mode property
